+++ Crafting/UI/RecipeFiltersCanvas.cs
@@ -1,9 +1,11 @@
-using BetterInventory.ItemSearch;
+using System;
+using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using SpikysLib.UI.Elements;
 using Terraria;
 using Terraria.Audio;
 using Terraria.DataStructures;
+using Terraria.GameContent.UI.Elements;
 using Terraria.ID;
 using Terraria.Localization;
 using Terraria.UI;
@@ -11,28 +13,46 @@ using Terraria.UI;
 namespace BetterInventory.Crafting.UI;
 
 public sealed class RecipeFiltersCanvas : UIState {
-    public UIFlexGrid filterList = null!;
+    public UIFlexList container = null!;
+    public UIPanel searchPanel = null!;
+    public UISearchBar searchBar = null!;
+    public UIFlexGrid filters = null!;
 
     public sealed override void OnInitialize() {
-        filterList = new() {
-            ListPadding = 6,
-            ItemWidth = 14
-        };
-        Append(filterList);
+        container = new() { ListPadding = 6 };
+        AddSearchBar(container);
+        AddFilters(container);
+        Append(container);
+    }
+
+    public override void Update(GameTime gameTime) {
+        base.Update(gameTime);
+        if (!filters.IsMouseHovering) Reflection.PlayerInput.MouseInModdedUI.GetValue().Remove("ModLoader/UIList");
     }
 
     protected override void DrawSelf(SpriteBatch spriteBatch) {
         base.DrawSelf(spriteBatch);
-        if (filterList.IsMouseHovering) Main.LocalPlayer.mouseInterface = true;
+        if (searchPanel.IsMouseHovering) Main.LocalPlayer.mouseInterface = true;
+        if (filters.IsMouseHovering) Main.LocalPlayer.mouseInterface = true;
+    }
+
+    public override void RecalculateChildren() {
+        base.RecalculateChildren();
+
+        if (searchPanel is not null) searchPanel.Width.Pixels = Math.Max(container.Width.Pixels, 14*4+3*6);
+    }
+
+    private void AddFilters(UIList container) {
+        filters = new() { ListPadding = 6 };
+        container.Add(filters);
     }
 
     public void RebuildRecipeGrid() {
         static void OnFilterChanges() {
-            if (Configs.BetterGuide.AvailableRecipes) Guide.FindGuideRecipes();
-            else Recipe.FindRecipes();
+            Utility.FindRecipes();
             SoundEngine.PlaySound(SoundID.MenuTick);
         }
-        filterList.Clear();
+        this.filters.Clear();
 
         EntryFilterer<Item, IRecipeFilter> filters = RecipeFiltering.LocalFilters.Filterer;
         for (int i = 0; i < filters.AvailableFilters.Count; i++) {
@@ -60,9 +80,55 @@ public sealed class RecipeFiltersCanvas : UIState {
                 filters.ToggleFilter(number);
                 OnFilterChanges();
             };
-            filterList.Add(item);
+            this.filters.Add(item);
         }
         Recalculate();
     }
 
+    private void AddSearchBar(UIList searchArea) {
+        searchPanel = new() {
+            Height = new StyleDimension(24f, 0),
+            BackgroundColor = new Color(35, 40, 83),
+            BorderColor = new Color(35, 40, 83)
+        };
+        searchPanel.SetPadding(0f);
+        searchArea.Add(searchPanel);
+
+        searchBar = new(Language.GetText("UI.PlayerNameSlot"), 0.8f) {
+            Width = new StyleDimension(0f, 1f),
+            Height = new StyleDimension(0f, 1f),
+            HAlign = 0f,
+            VAlign = 0.5f,
+            Left = new StyleDimension(-4, 0f),
+            IgnoresMouseInteraction = true
+        };
+
+        searchBar.OnStartTakingInput += () => searchPanel.BorderColor = Main.OurFavoriteColor;
+        searchBar.OnEndTakingInput += () => searchPanel.BorderColor = new Color(35, 40, 83);
+        searchPanel.OnLeftClick += (evt, _) => {
+            if (evt.Target.Parent != searchPanel) searchBar.ToggleTakingText();
+        };
+        searchBar.OnContentsChanged += OnSearchContentChange;
+        searchBar.OnCanceledTakingInput += () => OnSearchContentChange(null);
+        searchBar.SetContents(null, true);
+        searchPanel.Append(searchBar);
+
+        UIImageButton cancelButton = new(Main.Assets.Request<Texture2D>("Images/UI/SearchCancel")) {
+            HAlign = 1f,
+            VAlign = 0.5f,
+            Left = new StyleDimension(-2f, 0f)
+        };
+        cancelButton.OnMouseOver += (_, _) => SoundEngine.PlaySound(SoundID.MenuTick);
+        cancelButton.OnLeftClick += (_, _) => {
+            searchBar.SetContents(null, true);
+            SoundEngine.PlaySound(SoundID.MenuTick);
+        };
+        searchPanel.Append(cancelButton);
+    }
+
+    private void OnSearchContentChange(string? content) {
+        if (Main.gameMenu) return;
+        RecipeFiltering.LocalFilters.Filterer.SetSearchFilter(content);
+        Utility.FindRecipes();
+    }
 }